refactored to include emitter library
