+++ process_bigraph/core/processes.py
@@ -1,0 +1,7 @@
+"""
+==============
+Core Processes
+==============
+"""
+
+# TODO -- add core processes
