+++ process_bigraph/core/registry.py
