+++ process_bigraph/core/tests.py
@@ -4,9 +4,9 @@ Tests for Process Bigraph
 
 import random
 
-from process_bigraph.composite import Process, Step, Composite
-from process_bigraph.composite import merge_collections
-from process_bigraph.type_system import types
+from process_bigraph.core.composite import Process, Step, Composite
+from process_bigraph.core.composite import merge_collections
+from process_bigraph.core.type_system import types
 
 
 class IncreaseProcess(Process):
@@ -52,7 +52,7 @@ def test_process():
 
 
 def test_composite():
-    # TODO: add support for the various vivarium emitters
+    # TODO: add support for the various vivarium emitter
 
     # increase = IncreaseProcess({'rate': 0.3})
     # TODO: This is the config of the composite,
