+++ process_bigraph/experiments/minimal_gillespie.py
@@ -6,10 +6,9 @@ Note: This Process is primarily for testing multi-timestepping.
 variables and parameters are hard-coded. Do not use this as a
 general stochastic transcription.
 """
-import os
 import numpy as np
 
-from process_bigraph.composite import types, Step, Process, Composite
+from process_bigraph.core.composite import Step, Process, Composite
 
 
 class GillespieInterval(Step):
