--- process_bigraph/processes.py
@@ -1,9 +1,0 @@
-"""
-==============
-Core Processes
-==============
-"""
-from process_bigraph.composite import Process, Step, Composite
-from process_bigraph.registry import process_registry
-
-# TODO -- add core processes
