+++ process_bigraph/testing/test_emitter.py
@@ -1,9 +1,6 @@
 """Test the database emitter class with an instance used to emit and query simulation data."""
 
-
-import os
-
-from process_bigraph.emitter import DatabaseEmitter
+from process_bigraph.emitter.utils import DatabaseEmitter
 
 
 def run_simulation(time):
